pyproject.toml
[tool.poetry]
name = "llamaindex-course"
version = "0.1.0"
description = ""
authors = ["Raphael MANSUY <raphael.mansuy.partner@decathlon.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
llama-index = "^0.10.9"
python-dotenv = "^1.0.1"
transformers = {extras = ["torch"], version = "^4.38.1"}
chromadb = "^0.4.23"
llama-index-llms-ollama = "^0.1.2"
langchain = "^0.1.9"
langchain-community = "^0.0.24"
llama-index-embeddings-langchain = "^0.1.2"
scikit-learn = "^1.4.1.post1"
langchain-openai = "^0.0.7"
pydantic = "^2.6.2"
gradio = "^4.19.2"
sounddevice = "^0.4.6"
soundfile = "^0.12.1"
torchaudio = "^2.2.1"
langgraph = "^0.0.26"


[tool.poetry.group.dev.dependencies]
pylint = "^3.0.3"
pytest = "^8.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
